+++ src/NSchema/CliApplication.cs
@@ -2,6 +2,7 @@ using NSchema.Plan.PlanFile;
 using NSchema.Schema;
 using NSchema.Services.Reporting;
 using NSchema.State;
+using NSchema.State.Storage;
 
 namespace NSchema;
 
@@ -29,6 +30,9 @@ internal sealed class CliApplication(NSchemaApplication app, IConsoleMessenger m
 
     /// <inheritdoc cref="NSchemaApplication.PlanFile"/>
     public IPlanFileWriter PlanFile => app.PlanFile;
+
+    /// <inheritdoc cref="NSchemaApplication.State"/>
+    public ISchemaStateManager State => app.State;
 
     /// <inheritdoc cref="NSchemaApplication.Services"/>
     public IServiceProvider Services => app.Services;